refactor(PersonalizationForm): Extract status banner
